+++ packages/holz-env-filter/src/environment-filter.ts
@@ -1,26 +1,31 @@
 import type { LogProcessor } from '@holz/core';
 import { createPatternFilter } from '@holz/pattern-filter';
-import * as browserEnv from './browser-env';
-import * as serverEnv from './server-env';
+import { getEnvironmentVariable, getLocalStorageKey } from './get-env';
 
-export function createEnvironmentFilter({
+export const createEnvironmentFilter = ({
   processor,
   pattern,
   defaultPattern,
-}: Config): LogProcessor {
-  return createPatternFilter({
-    pattern: pattern ?? loadPattern() ?? defaultPattern ?? '*',
-    processor,
-  });
-}
+  localStorageKey = 'debug',
+  environmentVariable = 'DEBUG',
+}: Config): LogProcessor => {
+  /**
+   * Load from localStorage falling back to `process.env`. We always check
+   * both because we might be in Electron, where both are defined and equally
+   * valid.
+   */
+  const loadPatternFromEnvironment = () => {
+    return (
+      getLocalStorageKey(localStorageKey) ??
+      getEnvironmentVariable(environmentVariable)
+    );
+  };
 
-/**
- * Load from localStorage falling back to `process.env`. We always check both
- * because we might be in Electron, where both are defined and equally valid.
- */
-function loadPattern() {
-  return browserEnv.load(browserEnv.STORAGE_KEY) ?? serverEnv.load('DEBUG');
-}
+  return createPatternFilter({
+    processor,
+    pattern: pattern ?? loadPatternFromEnvironment() ?? defaultPattern ?? '*',
+  });
+};
 
 interface Config {
   /** Replace the detected pattern with something else. */
@@ -31,4 +36,10 @@ interface Config {
 
   /** Where to send the logs if they pass the filter. */
   processor: LogProcessor;
+
+  /** The localStorage key where patterns are saved. Default is `debug`. */
+  localStorageKey?: string;
+
+  /** The environment variable where patterns are saved. Default is `DEBUG`. */
+  environmentVariable?: string;
 }